fix parsing kernel versions with suffixes
cloudfoundry/guardian#232

When kernel versions had certain suffixes (e.g. `+`),
hashicorp/go-version would fail to parse the versions.

Instead, I took a simpler approach of just parsing the version manually.
The version is expected to have between 1 and 3 semver parts with no
prefix and any suffix. For version comparison, I took a similar approach
to the `KERNEL_VERSION` macro: https://github.com/torvalds/linux/blob/9d31d2338950293ec19d9b095fbaa9030899dcb4/tools/lib/bpf/bpf_helpers.h#L69

Effectively, it collapses the semver into a single number using the
formula: 2^16 * MAJOR + 2^8 * MINOR + PATCH

Rather than bounding the version parts to 8-bit, I figured it was safer
to use 16-bit to avoid overflow issues (which came up earlier this year
- see http://kroah.com/log/blog/2021/02/05/8-bits-are-enough-for-a-version-number-dot-dot-dot/)

+++ guardiancmd/seccomp_linux.go
@@ -5,7 +5,7 @@ import (
 	"github.com/opencontainers/runtime-spec/specs-go"
 )
 
-func MinPtraceKernelVersion() (int, int, int) {
+func MinPtraceKernelVersion() (uint16, uint16, uint16) {
 	return 4, 8, 0
 }
